move Cases and Case components to own directory, rename fetch to fetchRecords
